pyproject.toml
[project]
name = "tiktok-sentiment-analyzer"
version = "1.0.0"
description = "TikTok video downloader and AI-powered content analyzer"
requires-python = ">=3.8"
dependencies = [
    "yt-dlp>=2024.0.0",
    "openai-whisper>=20231117",
    "moviepy>=1.0.3",
    "pillow>=10.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

